Consolidate Concordance Search components
